+++ src/main/java/jimicloud/javagqlshowcase/model/Brand.java
@@ -1,10 +1,10 @@
 package jimicloud.javagqlshowcase.model;
 
-import org.springframework.data.annotation.Id;
-import org.springframework.data.relational.core.mapping.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Table;
 
 @Table
 @Data
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public final class Brand {
     @Id
-    private Integer id;
+    private Long id;
     private String name;
 }
