+++ src/main/java/jimicloud/javagqlshowcase/repo/BrandRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public interface BrandRepository extends ReactiveCrudRepository<Brand, Integer>{
+public interface BrandRepository extends ReactiveCrudRepository<Brand, Long> {
     @NotNull
-    Mono<Brand> findById(@NotNull Integer id);
+    Mono<Brand> findById(@NotNull Long id);
 
     @NotNull
     Flux<Brand> findAll();
