+++ src/main/java/jimicloud/javagqlshowcase/repo/ToolTypeRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public interface ToolTypeRepository extends ReactiveCrudRepository<ToolType, Integer> {
+public interface ToolTypeRepository extends ReactiveCrudRepository<ToolType, Long> {
     @NotNull
-    Mono<ToolType> findById(@NotNull Integer id);
+    Mono<ToolType> findById(@NotNull Long id);
 
     @NotNull
     Flux<ToolType> findAll();
